inner test class
